move errors to lib
