+++ pombot/lib/errors.py
